+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 2019-03-19
+
+- 新增 小程序音频中断开始和结束事件 wx.onAudioInterruptionBegin、wx.onAudioInterruptionEnd
+
 ## 2019-03-08
 
 - 更新`getUpdateManager Apis`的方法